Move myqsl_to_hive tests from test_operators
